Docs update

+++ 00-GUI_LIBRARY/utils/gui_string.c
@@ -261,7 +261,7 @@ uint8_t i;                                      //Number of bytes required for c
 //GUI_USE_UNICODE = 0: string length = 11; Length total: 11
 gui_string_prepare(&s, myStr);                  //Prepare string for reading
 while (gui_string_getch(&s, &ch, &i)) {         //Go through entire string
-    printf("I: %d, ch %c (%d)\r\n", i, ch, ch); //Print character by character
+    printf("I: %d, ch %c (%d)\r\n", (int)i, ch, (int)ch);   //Print character by character
 }
 \endcode
  * 
@@ -325,7 +325,7 @@ uint8_t i;                                      //Number of bytes required for c
 gui_string_prepare(&s, myStr);                  //Prepare string for reading
 gui_string_gotoend(&ptr);                       //Go to last character of string
 while (gui_string_getchreverse(&s, &ch, &i)) {  //Go through entire string
-    printf("I: %d, ch %c (%d)\r\n", i, ch, ch); //Print character by character
+    printf("I: %d, ch %c (%d)\r\n", (int)i, ch, (int)ch);   //Print character by character
 }
 \endcode
  *
